Removed scikit-learn dependence
the required code from scikit-learn is now included in mloop. When
scikit-learn releases the latest dev version, this code will be removed
and mloop will require scikit-learn as a dependency. But this is a
better solution than requiring installation of scikit-learn from source
in the short term.

Interface is no longer run as a seperate process but just a seperate
thread. This is to improve compatibility with matplotlib on MacOS and
Windows.