Add support for fitting the regularisation coeff
Every 20 fits, it generates a bunch of new networks with various
regularisation coefficients. It trains each on (most of) the current
data, and then tests on a cross validation set. If the best of these
performs significantly better than the current network on the cv set
then it replaces the current net with the new one. Disable this for now
because it seems like it still needs a bit of tweaking.

Also add a comment about possibly artifically capping costs.