Fixed one param visualization bug and typos in documentation
When optimizing one parameter, there were some issues reimporting the
saved files for the visualizations to work. This was due to the
problematic corner case of zero D or one D with one element arrays in
numpy. This has now been sanitized. Also fixed some critical typos in
the documentation.