Fix network saving/loading
Problem before was that I was using the same TF graph for everything, so
when making extra network it was adding the nodes to the graph rather
than using a new graph, and this confused the restoration (since the
checkpoint files only expected one net in the graph, but we had two).
Fixed this by associating a new tf.Graph with each net.

Also closed the TF session when we stop using a net. This is probably
good for efficiency.