Fixed bug where GaussianProcessVisualizer would error when plotting data from a learner archive that used a training archive which had different options set.
For example, if cost_has_noise was False during one optimization, then that optimization's learner archive was used as the training archive for another optimization, then the second optimization's archive would have issues. This occured because the number of fitted noise levels din't match the number of fits because the fits from the training run didn't fit a noise level.