Updated GaussianProcessLearner for parameter scaling
The Gaussian Process input params were not scaled using the sk StandardScaler as they are in the neural network. Also, the scipy optimize.minimize function does not work properly when the input parameters are large unless the inputs and bounds are scaled *before* the predictions. This is a problem also for the neural network learner, but has not been fixed by this commit