Fixed a matplotlib deprecation warning in `NeuralNetVisualizer.plot_density_surface()`.
Matplotlib wasn't happy about the interpolation method used by `pcolormesh()` given the dimensions of the data passed to it. Changing the interpolation to `'nearest'` fixes the issue. It's also the better way to perform the interpolation anyway since then the coloring of the squares is set by their value in their center rather than the value at one of their corners.