Record `best_in_extras` in controller archive
This PR modifies the `Controller` class to store the extras associated with the best run as an entry in the controller archive. For reference, the "extras" are any extra entries in the cost dict returned by the user's `Interface.get_next_cost_dict()` method beyond the standard entries interpretted by M-LOOP (i.e. other than `cost`, `uncer`, and `bad`).