Zero-index run numbers
Currently the learner archive's `best_index` is 0-indexed, as is typical for Python . However, the `best_index` in the controller archive is 1-indexed, so it differs from the value in the learner archive and is not the correct Python index. This occurs in part because M-LOOP counts runs (i.e. iterations) starting at 1 rather than 0. For better consistency, this PR modifies M-LOOP to zero-index runs. This also means that the `best_index` in the controller archive will now be zero-indexed.