Fix Nelder-Mead `init_simplex_corner` creation
As pointed out in #175, the way `self.init_simplex_corner` is created for `NelderMeadLearner` is problematic because it is set equal to `self.min_boundary`. This means that when `self.init_simplex_corner` is later modified, `self.min_boundary` is modified as well. This PR resolves the issue using the approach suggested in #175, namely by assigning a copy of `self.min_boundary` to `self.init_simplex_corner` rather than assigning the original array itself.