Remove erronenous text from `tutorials.rst`
As pointed out in #182, the "best parameters found" and "predicted best parameters" will generally differ, even in the absence of noise. This is because the "predicted best parameters" are not tested (they're what M-LOOP would try in a future iteration). However there is currently some text in `tutorials.rst` which indicates that those values should be the same. This PR removes that erroneous text from `tutorials.rst`.